requirements.txt
# filepath: c:\Users\Ivermectin\Music\Tanush\Try 3\requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.2
sqlalchemy==1.4.46
flask-migrate==3.1.0
flask-wtf==1.1.1
flask-login==0.6.2
mysql-connector-python==8.0.26
wtforms==2.3.3
email-validator==1.1.3
python-dotenv==0.19.0
werkzeug==2.2.3
flask-moment==1.0.2
pytest==7.3.1
pytest-flask==1.2.0
coverage==7.2.5
black==23.3.0
flake8==6.0.0
